Move gcm-logger.go from cmd/ to cmd/gcm-logger
When a Go package is installed via `go get` the binary name is assumed
to be the name of the package. This meant that gcm-logger was installed
as `cmd`. This patch moves it to a new subdirectory, `gcm-logger`, which
means `go get` will install it under a more descriptive name.